+++ packages/analytics/amplify_analytics_pinpoint_dart/lib/src/analytics_plugin_impl.dart
@@ -9,7 +9,6 @@ import 'package:amplify_analytics_pinpoint_dart/src/impl/analytics_client/endpoi
 import 'package:amplify_analytics_pinpoint_dart/src/impl/analytics_client/event_client/event_client.dart';
 import 'package:amplify_analytics_pinpoint_dart/src/impl/analytics_client/event_client/queued_item_store/dart_queued_item_store.dart';
 import 'package:amplify_analytics_pinpoint_dart/src/impl/analytics_client/session_manager.dart';
-import 'package:amplify_analytics_pinpoint_dart/src/impl/analytics_client/stoppable_timer.dart';
 import 'package:amplify_analytics_pinpoint_dart/src/impl/flutter_provider_interfaces/app_lifecycle_provider.dart';
 import 'package:amplify_analytics_pinpoint_dart/src/impl/flutter_provider_interfaces/cached_events_path_provider.dart';
 import 'package:amplify_analytics_pinpoint_dart/src/impl/flutter_provider_interfaces/device_context_info_provider.dart';
@@ -83,19 +82,17 @@ class AmplifyAnalyticsPinpointDart extends AnalyticsPluginInterface {
 
   @override
   Future<void> configure({
-    AmplifyConfig? config,
+    AmplifyOutputs? config,
     required AmplifyAuthProviderRepository authProviderRepo,
   }) async {
     // Parse config values from amplifyconfiguration.json
-    if (config == null ||
-        config.analytics == null ||
-        config.analytics?.awsPlugin == null) {
-      throw ConfigurationError('No Pinpoint plugin config available.');
+    if (config?.analytics?.amazonPinpoint == null) {
+      throw ConfigurationError('No analytics config available.');
     }
 
-    final pinpointConfig = config.analytics!.awsPlugin!;
-    final pinpointAppId = pinpointConfig.pinpointAnalytics.appId;
-    final region = pinpointConfig.pinpointAnalytics.region;
+    final pinpointConfig = config!.analytics!.amazonPinpoint!;
+    final pinpointAppId = pinpointConfig.appId;
+    final region = pinpointConfig.awsRegion;
 
     // Prepare PinpointClient
     final authProvider = authProviderRepo