+++ packages/auth/amplify_auth_cognito_test/test/plugin/auth_providers_test.dart
@@ -66,8 +66,9 @@ void main() {
 
     seedStorage(
       secureStorage,
-      userPoolKeys: CognitoUserPoolKeys(userPoolConfig),
-      identityPoolKeys: CognitoIdentityPoolKeys(identityPoolConfig),
+      userPoolKeys: CognitoUserPoolKeys(mockConfig.auth!.userPoolClientId!),
+      identityPoolKeys:
+          CognitoIdentityPoolKeys(mockConfig.auth!.identityPoolId!),
     );
 
     await plugin.addPlugin(authProviderRepo: testAuthRepo);