+++ packages/auth/amplify_auth_cognito_test/test/plugin/forget_device_test.dart
@@ -15,8 +15,9 @@ import 'package:test/test.dart';
 void main() {
   AmplifyLogger().logLevel = LogLevel.verbose;
 
-  final userPoolKeys = CognitoUserPoolKeys(userPoolConfig);
-  final identityPoolKeys = CognitoIdentityPoolKeys(identityPoolConfig);
+  final userPoolKeys = CognitoUserPoolKeys(mockConfig.auth!.userPoolClientId!);
+  final identityPoolKeys =
+      CognitoIdentityPoolKeys(mockConfig.auth!.identityPoolId!);
   final testAuthRepo = AmplifyAuthProviderRepository();
 
   late DeviceMetadataRepository repo;
@@ -31,7 +32,8 @@ void main() {
         secureStorage,
         userPoolKeys: userPoolKeys,
         identityPoolKeys: identityPoolKeys,
-        deviceKeys: CognitoDeviceKeys(userPoolConfig, username),
+        deviceKeys:
+            CognitoDeviceKeys(mockConfig.auth!.userPoolClientId!, username),
       );
       plugin = AmplifyAuthCognitoDart(
         secureStorageFactory: (_) => secureStorage,