+++ packages/auth/amplify_auth_cognito_test/test/state/sign_in_state_machine_test.dart
@@ -36,7 +36,7 @@ void main() {
     test('smoke test', () async {
       stateMachine
           .dispatch(
-            ConfigurationEvent.configure(userPoolOnlyConfig),
+            ConfigurationEvent.configure(mockConfigUserPoolOnly),
           )
           .ignore();
       await expectLater(
@@ -97,7 +97,7 @@ void main() {
       test('customAuthWithSrp requires password', () async {
         stateMachine
             .dispatch(
-              ConfigurationEvent.configure(userPoolOnlyConfig),
+              ConfigurationEvent.configure(mockConfigUserPoolOnly),
             )
             .ignore();
         await expectLater(
@@ -134,7 +134,7 @@ void main() {
       test('customAuthWithoutSrp forbids password', () async {
         stateMachine
             .dispatch(
-              ConfigurationEvent.configure(userPoolOnlyConfig),
+              ConfigurationEvent.configure(mockConfigUserPoolOnly),
             )
             .ignore();
         await expectLater(
@@ -186,7 +186,7 @@ void main() {
               ),
           completion(isA<Configured>()),
         );
-        deviceRepo = DeviceMetadataRepository(userPoolConfig, secureStorage);
+        deviceRepo = DeviceMetadataRepository(mockConfig.auth!, secureStorage);
         stateMachine.addInstance<DeviceMetadataRepository>(deviceRepo);
       });