+++ packages/notifications/push/amplify_push_notifications_pinpoint/test/pinpoint_provider_test.dart
@@ -5,7 +5,10 @@
 import 'package:amplify_analytics_pinpoint_dart/src/impl/analytics_client/analytics_client.dart';
 import 'package:amplify_analytics_pinpoint_dart/src/impl/analytics_client/endpoint_client/endpoint_client.dart';
 import 'package:amplify_analytics_pinpoint_dart/src/impl/analytics_client/event_client/event_client.dart';
+import 'package:amplify_core/src/config/amplify_outputs/notifications/amazon_pinpoint_channel.dart';
+import 'package:amplify_core/src/config/amplify_outputs/notifications/notifications_outputs.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
+import 'package:amplify_push_notifications_pinpoint/src/pinpoint_event_type_source.dart';
 import 'package:amplify_push_notifications_pinpoint/src/pinpoint_provider.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
@@ -30,9 +33,10 @@ void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
   final pinpointProvider = PinpointProvider();
   final mockAmplifyAuthProviderRepository = MockAmplifyAuthProviderRepository();
-  const notificationsPinpointConfig = NotificationsPinpointPluginConfig(
-    appId: 'APP_ID',
-    region: 'REGION',
+  const notificationsPinpointConfig = NotificationsOutputs(
+    amazonPinpointAppId: 'APP_ID',
+    awsRegion: 'REGION',
+    channels: [AmazonPinpointChannel.fcm, AmazonPinpointChannel.apns],
   );
   final awsIamAmplifyAuthProvider = MockAWSIamAmplifyAuthProvider();
   final mockAnalyticsClient = MockAnalyticsClient();
@@ -139,7 +143,7 @@ void main() {
       final properties = res.properties;
       final source = res.source;
       expect(properties.attributes.containsKey('journey_id'), isTrue);
-      expect(source, equals(PinpointEventSource.journey.name));
+      expect(source, equals(PinpointEventTypeSource.journey.name));
     });
 
     test(
@@ -151,7 +155,7 @@ void main() {
       final properties = res.properties;
       final source = res.source;
       expect(properties.attributes.containsKey('campaign_id'), isTrue);
-      expect(source, equals(PinpointEventSource.campaign.name));
+      expect(source, equals(PinpointEventTypeSource.campaign.name));
     });
   });
 
@@ -191,6 +195,32 @@ void main() {
           analyticsClient: mockAnalyticsClient,
         ),
         completes,
+      );
+    });
+
+    test('flush events timer initialized', () async {
+      when(
+        () => mockAmplifyAuthProviderRepository.getAuthProvider(
+          APIAuthorizationType.iam.authProviderToken,
+        ),
+      ).thenReturn(awsIamAmplifyAuthProvider);
+      when(
+        () => mockAnalyticsClient.init(
+          pinpointAppId: any(named: 'pinpointAppId'),
+          region: any(named: 'region'),
+          authProvider: any(named: 'authProvider'),
+        ),
+      ).thenAnswer((realInvocation) async {});
+
+      await pinpointProvider.init(
+        config: notificationsPinpointConfig,
+        authProviderRepo: mockAmplifyAuthProviderRepository,
+        analyticsClient: mockAnalyticsClient,
+      );
+
+      expect(
+        pinpointProvider.autoEventSubmitter.duration,
+        const Duration(seconds: 10),
       );
     });
 
@@ -316,6 +346,48 @@ void main() {
       verify(mockEndpointClient.updateEndpoint);
     });
 
+    test('registerDevice should run successfully when token is expired',
+        () async {
+      when(
+        () => mockAmplifyAuthProviderRepository.getAuthProvider(
+          APIAuthorizationType.iam.authProviderToken,
+        ),
+      ).thenReturn(awsIamAmplifyAuthProvider);
+      when(
+        () => mockAnalyticsClient.init(
+          pinpointAppId: any(named: 'pinpointAppId'),
+          region: any(named: 'region'),
+          authProvider: any(named: 'authProvider'),
+        ),
+      ).thenAnswer((realInvocation) async {});
+
+      final mockEndpointClient = MockEndpointClient();
+
+      when(mockEndpointClient.updateEndpoint)
+          .thenThrow(const UnknownException('message'));
+
+      when(
+        () => mockAnalyticsClient.endpointClient,
+      ).thenReturn(mockEndpointClient);
+
+      await expectLater(
+        pinpointProvider.init(
+          config: notificationsPinpointConfig,
+          authProviderRepo: mockAmplifyAuthProviderRepository,
+          analyticsClient: mockAnalyticsClient,
+        ),
+        completes,
+      );
+
+      expect(
+        pinpointProvider.registerDevice(
+          '',
+        ),
+        completes,
+      );
+      verify(mockEndpointClient.updateEndpoint);
+    });
+
     test('recordEvent should run successfully', () async {
       when(
         () => mockAmplifyAuthProviderRepository.getAuthProvider(
@@ -363,7 +435,7 @@ void main() {
       verify(
         () => mockEventClient.recordEvent(
           eventType:
-              '${PinpointEventSource.campaign.name}.${PinpointEventType.foregroundMessageReceived.name}',
+              '${PinpointEventTypeSource.campaign.name}.${PinpointEventType.foregroundMessageReceived.name}',
           properties: any(named: 'properties'),
         ),
       );