+++ packages/amplify_datastore_plugin_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.3.0
+
+- Minor bug fixes and improvements
+
+## 2.2.0
+
+- Minor bug fixes and improvements
+
+## 2.1.0
+
+- Minor bug fixes and improvements
+
 ## 2.0.0
 
 - Minor bug fixes and improvements