+++ packages/common/amplify_db_common_dart/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.4.4
+
+- Minor bug fixes and improvements
+
+## 0.4.3
+
+- Minor bug fixes and improvements
+
+## 0.4.2
+
+- Minor bug fixes and improvements
+
 ## 0.4.1
 
 ### Fixes